+++ NEWS.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] 02/12/2025
+
+### Added
+
+* `ApproximateExponentialRetraction` and `ApproximateLogarithmicInverseRetraction` as types
+  to specify approximation methods for exp and log, which due to their approximation are also
+  only approximate retractions and inverse retractions, respectively. (#248)
+
 ## [2.2.1] 13/11/2025
 
 ### Added